+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## Next Version
+
+### Features Added
+- Migrate jackson `ObjectMapper` instances to `JsonMapper` and `YamlMapper` builders to resolve deprecation warnings [#507](https://github.com/ConsenSys/web3signer/pull/507)
+
 ### Bugs Fixed
 - Upgrade Vertx to 4.x, signers to 2.0.0 and various other dependencies to latest versions [#503](https://github.com/ConsenSys/web3signer/pull/503)